Added markers for GBIF data
The map now renders markers for GBIF data as a different color while still responding to all buffer changes. Updated the buffer text to specify kilometers, added table headers that dynamically change.

Next steps are to allow the user to specify how many records they want to receive and to add a loading circle for when GBIF is queried to communicate the load time.